pyproject.toml
[project]
name = "letswork-autofil"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "holidays>=0.75",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.0",
    "webdriver-manager>=4.0.2",
    "workadays>=2023.12.30",
]
